[UPD] Adicionando validação ao excluir turma

+++ excluir.php
@@ -3,9 +3,10 @@
     echo "<br>";
 
     $oTurmaDAO = new TurmaDAO();
-    $oTurmaDAO->delete($_POST['id']);
+    if ($_POST['id']) {
+        $oTurmaDAO->delete($_POST['id']);
+    }
 
     header('Location: turma.php');
 
     require_once "inc/Footer.php";
-
